Expand the script and the description
Expand the script to take a commandline argument for the output file,
but default to stdout. Update the description accordingly. Add some
bells and whistles to the script to make it more user friendly, such as
printing a help text and checking whether the shrunken output is
actually smaller than the input.