shellcheck: avoid unnecessary echo
Shellcheck advises:

  SC2005: Useless echo? Instead of 'echo $(cmd)', just use 'cmd'.

The extra echo was added in 9bd52f as a workaround for `wc' in OSX,
which printed extra whitespace before the character count. Fix this
issue in a different way by piping the `wc' output through `awk'.

This also means that we no longer use the suggested workaround, so
remove the dangling acknowledgement.