#5: replace optional positional arguments with option flags
The script currently takes up to three positional arguments, with one of them
required and the other two optional. It also takes two option flags.

This is a bit mix-and-match, and will not scale if we add new features in the
future. Replace it with option handling through `getopts' for the optional
arguments.

Resolves #5.