Add inplace option -i to replace the input
Add an option -i to replace the input file with the shrunk output (if
smaller). This creates a temporary directory in a secure and portable
manner, cleaning it up afterwards.

The temporary directory is created using 'mktemp' if present, otherwise
the less secure method of including the PID, current time and possibly
(if available) $RANDOM is used. The directory is removed upon exit, but
carefully: 'rm -fr' is only executed on a then-empty directory after
output.pdf has been removed. This could leak directories, but I consider
that less of a problem than possibly removing something important.